[hotfix #129] 앨범 로직 수정 (#130)
* [hotfix #129] Album Entity 연관관계 albumDesign으로 수정

* [hotfix #129] response AlbumCover에서 AlbumDesign으로 수정

* [hotfix #129] column명 수정

* [hotfix #129] albumCover타입 Long으로 변경

* [hotfix #129] 회원정보 입력 API 로직 수정
- albumCover 추가하지 않고 앨범 추가 시 design 업데이트하도록 수정

* AUTH_WHITELIST 추가

+++ src/main/java/com/pophory/pophoryserver/domain/member/dto/request/MemberCreateRequestDto.java
@@ -27,5 +27,5 @@ public class MemberCreateRequestDto {
     private String nickname;
 
     @Schema(description = "앨범 커버 1 ~ 4 중의 값으로 요청", example = "1")
-    private int albumCover;
+    private Long albumCover;
 }
